backend/requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
tensorflow
Pillow
markupsafe==2.0.1
